[daemon utils] move platform/check_daemon_status.py to common/platform/daemon_utils.py
