+++ tests/common/platform/daemon_utils.py
